refactor: move header to components
